+++ interface/app/config_widgets.py
@@ -33,6 +33,10 @@ class ConfigWidgetsMixin:
         except Exception:
             pass
         try:
+            self._box_refill_var.set(bool(puzzle.get('box_refill_enabled', False)))
+        except Exception:
+            pass
+        try:
             self._opacity_slider.set(puzzle['overlay_opacity'])
             self._refresh_opacity_value()
         except Exception: